requirements.txt
numpy
pillow
imageio
scipy
tkinter